+++ tests/test_code_gen.py
@@ -1,3 +1,4 @@
+# coding: utf-8
 """
 Part of the astor library for Python AST manipulation
 
@@ -663,6 +664,19 @@ class CodegenTestCase(unittest.TestCase, Comparisons):
         x = f'{f.name}={{self.{f.name}!r}}'
         '''
         self.assertSrcRoundtripsGtVer(source, (3, 6))
+
+    @unittest.skipUnless(sys.version_info >= (3, 8, 0, "alpha", 4),
+                         "f-string debugging introduced in Python 3.8")
+    def test_fstring_debugging(self):
+        source = """
+        x = f'{5=}'
+        y = f'{5=!r}'
+        z = f'{3*x+15=}'
+        f'{x=:}'
+        f'{x=:.2f}'
+        f'alpha α {pi=} ω omega'
+        """
+        self.assertAstRoundtripsGtVer(source, (3, 8))
 
     def test_docstring_function(self):
         source = '''